notifier: make sure generated notices always have a backtrace
In theory it's possible to generate a backtrace that consists only of
internal frames. When this happens, then we send a notice without a
backtrace, because all the internal frames are filtered. When we see a
notice without a backtrace in the dashboard, it's incredibly hard to
debug it.

This change ensures that in case of an internal error we at least can
debug the failure.