filters: make it possible to filter out user
A user reported that it's not possible to filter context/user. It's a
reasonable request given that the user information has been moved to
context recently.

Here we special-case context/user because we do want to be able to
filter it, but we don't want to filter out other payload in `context`.