notice: fix infinite loop when truncating payload
Fixes #82 (Airbrake-ruby stuck in an infinite loop)

This is the first step to fix the problem. Firstly, we must identify
the bad payload piece, then we will be able to decide what to do next.

This PR at least fixes the infinite loop. Apparently we forget to
filter something, so we reduce truncator's `max_size` and eventually
it becomes `0`. This is a sign things went wrong, so we can return
early.