notice: deprecate context/action compatibility code
This was needed to ease transition from v4 to v5, but now we don't
want to support this special case due to the ability to `build_notice`
and assign `component/action` to the notice directly. It was also
impossible to have `component/action` as a normal param key.

Deprecating it first, so it can be removed in the next release. I wasn't
careful enough, so some of our integrations use this legacy API.