airbrake-ruby,notifier: use the Null Object pattern for notifiers
Fixes https://github.com/airbrake/airbrake/issues/713
(undefined method `build_notice' for nil:NilClass)

It becomes really tedious to keep track of whether a notifier is
configured or not. Since we introduced `Airbrake.[]` to retrieve
notifiers, all integrations have to add an additional check for it:

```rb
if notifier = Airbrake[:default]
  notifier.notify(...)
end
```

With the current change this check is not needed anymore.