Enable truncation of frozen hashes
Currently when truncating hashes for logging purposes, the original hash values get modified.

While immutability would be preferred, reusing the hash is likely here for keeping memory usage down. This is however creating an error when a frozen hash is passed in.

This change resolves this by checking for a frozen hash, and duplicates it allowing for modification when required.