Gemfile: lock rubocop to v0.51.0
Every time Circle runs a build, it pulls the newest Rubocop. This causes
troubles for contributors and personally me. Instead of working on new features
or fixes we have to work on fixing Rubocop offences prior to merging actual
useful PRs. This happens because we used to specify an open-ended version
pattern for Rubocop in our Gemfile which was deliberate. Now it's obvious that
this approach isn't good.

With this change I lock Rubocop to a specific version, so it doesn't interrupt
anyone. I'll be updating Rubocop manually from time to time and fix offences in
a sweep.