async_sender: prevent race conditions in #has_workers?
Fixes #312 (TypeError: can't move to the enclosed thread group)

`@workers.list.empty?` is racey, so if multiple threads are spawned at
the same time, one of them would spawn workers, the other one would
see that the list is empty, too and try to spawn the workers again.

I have to admit that storing the workers in a ThreadGroup instead of
an Array was a damn good choice. Luckily, it saved us from the worker
leak disaster.