Implement basic API for merging context
Fixes #314 (Possible to set context for possible error?)

This feature has been request many times already. While Airbrake supports it,
it's not obvious for users how to make it work. It's also hard to advertise it
since there's no simple method to call to explain the API. Previously, I used to
link people to
https://github.com/airbrake/airbrake/issues/696#issuecomment-366857391

I guess it is not very convenient to use that API. Therefore, the feature
doesn't use Threads and is much simpler to use. Users no longer need to mess
with threads and cleaning up context information is on us.