notifier: yield notice before filter_chain is invoked
The order switch is needed to make inline filters usable in the ignore scenario.

Let's say we set notice's component in the inline filter. Then we define a
filter to ignore it via `Airbrake.add_filter`.

Before: `Airbrake.add_filter` filters are invoked. The inline filter is invoked
next. Nothing is ignored.

After: The inline filter is invoked. `Airbrake.add_filter` filters are reading
that and ignoring the notice.

This is exactly the problem with our Logger integration (it appears to be
impossible to ignore logger messages).