notifier: implement #inspect and #pretty_print
It has always bothered me that when you configure a notifier in Pry, it would
print a huge object state, which is barely readable. Quite often I would need to
exit pager with (q) to be able to type again.

With help of this change we get a meaningful inspect output, so only relevant
information is shown.

The default filter stuff was also moved from Notifier to FilterChain because
Rubocop was complaining that the Notifier class is too big.