route_sender: replace min/max with TDigests
https://github.com/tdunning/t-digest

> A new data structure for accurate on-line accumulation of rank-based
> statistics such as quantiles and trimmed means. The t-digest algorithm is also
> very parallel friendly making it useful in map-reduce and parallel streaming
> applications.

This introduces a few new dependencies:

tdigest (https://github.com/castle/tdigest)
|__rbtree (https://rubygems.org/gems/rbtree)

I am locking the `tdigest` dependency because it doesn't seem to follow
semver. It also is not in active development, so a sudden minor bump may break
Airbrake Ruby for a lot of customers. This is especially true since we
monkey-patch the class to pack tdigests how we need.