Drop support for Ruby 2.0
Our newest dependency, `tdigest`, depends on `rbtree`. RBTree fails to compile
on 2.0.0p648: https://circleci.com/gh/airbrake/airbrake-ruby/2164

```
checking for rb_exec_recursive() in ruby.h... *** extconf.rb failed ***
  ...
/usr/local/lib/ruby/2.0.0/mkmf.rb:434:in `try_do': The compiler failed to
generate an executable file. (RuntimeError)
You have to install development tools first.
```

There are some patches on the internet that fix RBTree mentioned in:
https://github.com/seki/Drip/issues/4

There's also RBTree 2, which may fix the problem:
https://github.com/kitak/rbtree2

That said, we don't control what `tdigest` depends on, so we don't have any
means to experiment.

Now, I'm usually reluctant to dropping Rubies without real need. In theory if we
controlled `tdigest`, we could still continue supporting Ruby 2.0. It's possible
to fork it and hack on it. However, I don't think it's the right approach since
Ruby 2.0 is almost dead. It is simply not worth bothering with this because 99%
of users already migrated to newer Rubies.