route_sender: fix possible race with increment_stats
Sometimes we have routes that report 0 count. In theory, this should never
happen because if there's a route record, then the count should be at least 1.

I suspect this happens due to a possible race between `send` and
`increment_stats`.

This gets executed: `@routes[route] ||= RouteStat.new`

Then the scheduler switches to `send` and sends the routes. This essentially
sends the route with zero data.

So far I have no better idea and this is kind of a guess rather than a solid
proof.