Rename Notifier to NoticeNotifier
Now that we have all sorts of notifiers in this library, the name `Notifier`
doesn't look too descriptive. It's too generic. It was tempting to rename it to
`ErrorNotifier` but this would likely be confusing. Although we send errors to
Airbrake, the object that we work with is called `Notice`. Had we used `Error`,
`ErrorNotifier` would make more sense, but alas.

Accessing `Airbrake::Notifier` is still possible but each time doing so emits a
warning. Although `Airbrake::Notifier` was marked as private API since its
creation, it's too tempting to use this class in your codebase and I suspect
some people might do that, so I am just being cautions here.