airbrake-ruby: validate config in Airbrake#configure
This change aims to remove duplicative code from notifiers. Every notifier used
to check if the config is valid and raise an error. It makes sense to avoid this
duplication by moving config validation on the layer above, closer to the public
API.

As result, notifiers stopped accepting hashes and expect an instance of Config.
In case people initialize notifiers directly (shouldn't be the case, but who
knows), we show a deprecation warning, that passing a hash is not supported
anymore.