Implement Airbrake.notifiers
This new method returns all configured notifiers. This way it's possible to
access performance and deploy notifiers, which is going to be used in the
airbrake gem.

The old `Airbrake#[]` method is a weird citizen now. It can retreive only notice
notifiers. I am not deprecating it yet but in the future it might be removed or
merged with `Airbarke.notifiers`.