Build this gem on JRuby properly
The current approach doesn't work with JRuby. The patch that aimed to fix the
issue didn't actually work (duh):
https://github.com/airbrake/airbrake-ruby/pull/408

The problem is that I run `gem build *.gemspec` using CRuby, so the JRuby check
inside the gemspec always has the same outcome. This change builds gems
properly, with `rake gems`.