Factor out Logger to the Loggable module
Since the Airbrake module holds only one instance of every notifier, it opens
the gate to centralised logging. We no longer need to toss logger instances
around and inject them into every class that needs logging.

It always bothered me that we were doing it but it was the only way to go if we
wanted to maintain a fat wrapper that the Airbrake module was. Loggable is super
easy to use and I was able to `include` and it "just worked". Moreover, testing
is a ton easier now.