Use one Config instance throughout the code
This is a huge win that became possible ever since we dropped support for
multiple notifier configuration. We no longer need to inject config to every
single class that wants to read a value. Instead, we simply access the instance
through `Airbrake::Config.instance`.

This should still be thread-safe because the config is initialized in the
beginning and during the program's lifetime we only do reads and never ever
overwrite.