performance_notifier: attach 'environment' to the top level of JSON
There was a change in the backend: instead of sending environment information
with every request/query, we send it at the top level of the JSON payload.

This is a breaking change but the airbrake gem hasn't been updated yet to work
with environments, so it's not a huge deal.