Use Config#check_configuration in Notifiers
We want to be able to add_filter's to Airbrake even when it's not
"configured". This resolves the headache with load order in the Rails
integration: add whenever you can but notify only when configured.

It also resolves the problem with notify methods that return nil. Users of
our API had to perform annoying nil checking, now they don't need this and can
fully rely on promises.