performance_notifier: make resource name independent
Currently, we require that endpoint destination should match the resource
name. This is rather convenient since we can rely on convention, however if the
backend doesn't follow it, our code breaks (if we add new models).

With this change we use the cargo->destination metaphor and add it to our
models, so every model declares what it carries and where it should be
sent. PerformanceNotifier doesn't make any assumptions anymore.