airbrake-ruby: make sure optional filters are loaded
Fixes #460 (blacklist_keys config doesn't work)

The problem was that we never reconfigured already instantiated notifiers. We
always spawn notifiers as soon as the library is required but they get spawned
without some of default filters that are supposed to work given certain options
such as `blacklist_keys`.

The new `#reset` method respawns all notifiers on `#configure`. We have to
respawn so the new notifiers pickup the config. We need to spawn without the
user config so all the methods are working (e.g. notify returns a rejected
promise and such).