Add Airbrake::Benchmark
This module provides methods for measuring performance of Ruby code. Its
`.measure` returns monotonic time of elapsed operation.

This will be useful in the airbrake gem where we need to calculate time spent on
executing HTTP requests. In the future it may come in handy for other tasks as
well.