airbrake-ruby: initialize notice notifier with filters
Fixes #468 (blacklist_keys not being filtered out in version 4.2.3)

The problem was that at the time when we initialize NoticeNotifier (on library
load) the Config has no blacklist/whitelist keys specified. Therefore, when we
`configure` Airbrake, newly set values have no effect.

With this change we move conditional filter initialization out of
NoticeNotifier. The risk is that calling `configure` twice will append the same
conditional filters twice. Therefore, `reset` must be called prior to second
`configure` (but it's probably a good idea not to call it twice and it should be
very uncommon).