performance_notifier: always join the background thread
In the current version of the code the background thread clears the variable
that the parent thread reserved for it. This is illogical and probably error
prone.

With this change the parent thread takes care of cleaning up the child, so that
the child knows nothing of how it is managed. This makes much more sense.