performance_notifier: reuse the same thread for flushing
Previously, we would spawn a new thread every 15 seconds. According to
RackMiniProfile, this would use around 1.05MB of memory per request on my route
that makes a few SQL queries, which makes it #1 gem that wastes memory for my
Rails 5 app (both allocated and retained).

Spawning threads implies significant overhead, so I realised we don't actually
need to spawn anything. We can have just one thread that sleeps until more work
is given. Then this timer thread would fire and sleep again (unless there's more
work to do).

As result, we use a lot less memory (basically, only the memory needed to
pass payload to the worker thread). The downside is that we no longer support
promises properly (if a request fails in background, the promise will not tell),
but this seems to be a reasonable compromise for the benefit we gain.