performance_notifier: use thread pool to send APM data
Currently, for sending APM data we use just one timer. It is responsible for
everything: accumulating data, waiting and sending it. This is rather slow
because the timer usually makes 3 subsequent HTTP requests: one for performance
breakdown data, one for request data and one for query data. More so, it has
only 15 seconds to do that. This is because our flushing interval is 15 seconds
by default. If the requests stall, we risk losing subsequent APM data.

Instead, we will use AsyncSender to send APM data. This way PerformanceNotifier
will have its own independent thread pool (from NoticeNotifier) that will be
processing all the APM-related jobs. This is more fault-tolerant: the timer
doesn't need to wait for HTTP requests to complete, it just tells AsyncSender
what to send and becomes ready for another flush.