performance_notifier: add #notify_sync
`PerformanceNotifier#notify_sync` exists for the same reasons as
`NoticeNotifier#notify_sync`.

In certain scenarios, such as Resque jobs, `AsyncSender` dies before it can send
something. Because `PerformanceNotifier` waits 15 seconds before sending
payload, it's guaranteed that once a Resque job finishes, it also kills
`PerformanceNotifier` before it can send. This is the unfortunate effect of
forking.

To cope with this, we must send synchronously.