notice_notifier: fix confusing error message
Addresses #524 (Error Airbrake: performance notifier closed causes deadlock)

A user reported that he saw the following message in the logs:

> Airbrake::Error (attempted to build deadlock; recursive locking with closed
  Airbrake instance)

Because exceptions are stringified, the error message can get very confusing. In
issue #524 the user thought that there's a deadlock in airbrake-ruby (me too,
to be honest), but when I took a closer look, it turned out that we print the
exception we're building as a string.

I changed the message, so it's more obvious what's going on.