Converted all the Struct with keywords classes to classes
The classes here that are derived from Struct all are initialized with
keyword parameters. Ruby 2.7 adds built-in support for this kind of
thing with a keyword_init option, but the new Struct is not compatible
with the old behavior in this case, and generates a lot of warnings if
the keyword_init option isn't specified.

To avoid all that, just converted them to regular classes.

[Fixes #536]