Configuration fixes for airbrake/airbrake#1062
Airbrake::Rack::Instrumentable#airbrake_capture_timing does method
wrapping similar to alias_method_chain, but this style of wrapping
will cause stack overflows on methods that have been overriden
with Module#prepend. However, using prepend style wrapping will
stack overflow on methods built with alias_method_chain or the
equivalent alias_method calls.

Some people have chained codebases, like those based on Rails < 5.
Rails 5+ codebases are based on prepend. For either case, there
will be random methods that are implemted in the other style, so
even if we set a default we need to have the option of using the
other style. We need to support both wrapping methods, and we
need a way for the user to declare which wrapping method is used
as the default for their project. It can't be autodetected.

Method wrapping also needs to do argument forwarding, and the
syntax and semantics for that changed in Ruby 2.7, effectively.
The real change is coming in Ruby 3 but 2.7 adds many warning
messages for the transition.

Here, we just add the config setting for the default method
wrapping style, and a flag constant to indicate whether the
new argument forwarding style and syntax is supported. As new
Ruby platforms support the new syntax, the flag constant can
be adjusted to match.

The rest will be done in the airbrake gem.