async_sender: fix handling of performance stats on full backlog
There's a bug in the async sender that can be reproduced under the following
circumstances:

1. AsyncSender's backlog must be full
2. Next message is a performance payload stat

The code expects that we pass a notice (a hash version of it) and therefore it
expects that `notice[:errors]` must exist. This is not the case for performance
stat payload, which doesn't contain `:errors`.

Therefore, when we attempt to print the message in the `will_not_deliver`
method, it crashes with the following error:

```
/lib/airbrake-ruby/async_sender.rb:63:in `will_not_deliver':
undefined method `first' for nil:NilClass (NoMethodError)
```

While thinking about the solution, I figured that we shouldn't print that inside
`AsyncWorker` since it's really `ThreadPool`'s concern.

`AsyncWorker` is still responsible for rejecting the promise, though.