filters/keys: add referer, httpMethod, userAddr & userAgent
In the Airbrake gem we have filters that add the following keys to the context
payload:

* referer
* httpMethod
* userAddr
* userAgent

This information is not critical for our services to work and sometimes users
need to filter that due to various reasons. Specifically, someone requested to
be able to filter `userAddr` due to GDPR reasons.