airbrake-ruby: do not reappend same filters
Fixes #561 (Multiple `Airbrake.configure` calls append the same filters to the
filter chain)

This solution makes sure we don't reappend the same "static" filters over and
over when we call `Airbrake.configure`. Static here means that the filters
implement exactly the same algorithm, so they are duplicative and thus not
necessary.

There are non-static filters. For example, if users configure blacklist keys
twice, the same filter will be added. The filter chain will have two instances
of `KeysBlacklistFilter`. However this is okay because both of them will filter
different keys (therefore the algorithm is different).