filters/git_last_checkout: add support for AIRBRAKE_DEPLOY_USERNAME
In certain environments, when a deploy starts, end deploy user might not be the
one who started the deploy. For example, if you deploy via Ansible, the script
could create a fake user who would pull the web app repo and deploy it for
you. Even though you started the script, the filter will pick up the fake user
name, since `.git` was created by that user.

To work around that, we introduce the `AIRBRAKE_DEPLOY_USERNAME` env
variable. With help of it we can forward the username from Ansible to Ruby.