thread_pool: create a new thread group per process
Fixes #576 (ThreadError: can't move to the enclosed thread group)

When we `Airbrake.notify` inside a `fork` block, `ThreadPool` tries to respawn
workers (which were killed by the `fork`) to the same `ThreadGroup`, which was
created in the parent process. The problem is that that `ThreadGroup` is already
finalised in that parent process and no new threads can be added anymore to the
group. The fix is to create a new `ThreadGroup` on `fork`. This way every
process will have its own `ThreadGroup`.