Make it possible to configure remote config host
At the moment there's a bug with remote configuration. A user would install the
new notifier but the notifier would use our staging S3 bucket to fetch
configs. Instead, we should hardcode the bucket path to production because
that's the goal. A helper option called `remote_config_host` was added. With
help of it can easily test the functionality.