remote_settings: do not turn on options disabled by the config
There's a bug where performance stats or error notifications cannot be disabled
through the config:

```rb
Airbrake.configure { |c| c.performance_stats = false }
```

Once that code is loaded, remote configuration will be used and APM will be
enabled again, when users don't want that.

The fix is to remember the original setting and when it's false, we shouldn't
ever enable it.