remote_settings: delete the code for config dumping/loading
Some of our users complain that our airbrake-ruby notifier tries to dump the
remote config that we pull from S3 to the gem directory. The OS user that is
running the code may not have write access to that directory and therefore
writing the remote config will fail.

There's no good universal location to keep persistent remote config, so it's
easier just to delete the config.