Drop support for Ruby 2.3 & 2.4
We no longer test on these rubies and this caused some [trouble][1] with the
Airbrake gem. We didn't catch this error becasue the Airbrake gem still tests on
2.3 & 2.4, but airbrake-ruby doesn't. We also never dropped support 2.3 & 2.4,
which left the whole situation in limbo.

```
5) Airbrake::Rack::Instrumentable.********_capture_timing when request store has a route attaches timing for a method with all arg types
     Failure/Error: params[:groups].merge!(timed_trace.spans)

     NoMethodError:
       undefined method `transform_values' for #<Hash:0x000055b7ee9f7300>
```

Since 2.3 & 2.4 are EOL for a long time (March 31, 2020 for 2.4), we can safely
stop supporting these rubies.

[1]: https://app.circleci.com/pipelines/github/airbrake/airbrake/215/workflows/d3f04866-05b9-4da6-ba15-b0c285eb0c23/jobs/9590