thread_pool: change severity of "full queue" msg from error to info
Some of our users complained that `message.inspect` "pressures" their server
memory. Indeed, the dumped string can be large (up to 64KB).

Instead, let's change the severity of this message to `info`, so that it won't
be printed by default. It's not really an error anyway, and the `info` severity
suits it perfectly.

`message.inspect` also won't be called by default anymore, since it's wrapped in
a block now.