Add Backlog
The Backlog class accepts notices and APM events and synchronously sends them in
the background at regular intervals (2 minutes).

The notices that it accepts are those that have failed to be sent due to network
issues or Airbrake servers struggling. If the second attempt fails, then such
data is discarded forever.

The maximum backlog size is 100. If we cannot add to the backlog anymore, the
item would be skipped and an error would be logged.