Mock the git logs instead of using real ones (#716)
The specs were using the real .git/logs/HEAD file and it's contents.
This was causing a failure when run in the context of our github actions
CI builds. This update mocks the access to the file and it's contents.

We did consider mocking the filesystem but this was simpler and didn't
introduce an additional dependency.

Paired with @thompiler on this one.