Do not enclose the worker ThreadGroup (#714)
Enclosing the ThreadGroup has caused issues for anyone using
version 0.3.1 of the timeout gem. @thompiler and I were able to
replicate the issue by running a rails 7.0.4 app with ruby version 3.2.1
and the timeout gem at version 0.3.1 `timeout, '= 0.3.1'` in the
Gemfile. It was also important to run the rails server with
`WEB_CONCURRENCY=3`.

```
WEB_CONCURRENCY=3 bin/rails s
```
Then watch the airbrake logs you may need to enable this in your
airbrake.rb initializer. Any Net::HTTP request will trigger a timeout on
the enclosed ThreadGroup and cause an issue. This could be sending an
Airbrake notice, performance data or fetching the remote config.